🎨 Split the Command class into an AbstractCommand class
